requirements.txt
appdirs==1.4.4
asgiref==3.5.2
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cssselect==1.2.0
Deprecated==1.2.13
Django==4.1.4
gunicorn==20.1.0
howdoi==2.0.20
idna==3.4
keep==2.10.1
lxml==4.9.2
Pillow==9.3.0
pycparser==2.21
PyGithub==1.57
Pygments==2.13.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyquery==1.4.3
requests==2.28.1
rich==12.6.0
sqlparse==0.4.3
terminaltables==3.1.10
tzdata==2022.7
urllib3==1.26.13
wrapt==1.14.1
